Added Dockerfile for release image, rename existing to development
The existing Dockerfile was meant for development purposes only. It is
still available to use with `docker-compose` to develop on the backend.

However, the new release file builds a Docker image containing all PHP
modules *and* the actual backend itself. This way, it can be directly
ran provided the database credentials are supplied to it.