Adjusted logic for invalid filenames in post_submit_update_file.php
Responses to the app remain the same as before, but now if we know the filename has ".zip" in it, but doesn't include "Oxygen", we continue processing so that the filename is saved to the DB & Discord webhook notifications are sent as well.

This was done because the "Oxygen" check is no longer a guarantee that the filename is a valid OnePlus OTA filename. Example: 8-series Open Beta 11 on 17th June 2021 switched to ota-manual-sg.allawnofs.com where files are named as hyphened GUIDs.
This could have been easily incorporated into validity checks (using regex) but we need to be aware of any filename changes OnePlus may have implemented.